Updates to instructions.html
This modifies the ordered list under the Setup Instructions heading to
better reflect the file/directory structure that will exist on the
USB drive used by Contributor Day participants. The `vvv-contribute.box`
file does not actually get included in `vvv.zip`.

This also updates the Logging In section to reflect that the Meta
Environment is included as well.